Data Science Projects/1.goodreadsbookanalysis/requirements.txt
# -*- coding: utf-8 -*-

# Python libraries for Goodreads Book Analysis
beautifulsoup4==4.12.3
requests==2.31.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
pillow==10.3.0

# Additional libraries for Jupyter and visualization
ipykernel==6.29.4
jupyterlab==4.1.8

